requirements.txt
Flask==2.0.1
Werkzeug==2.0.1
opencv-python==4.5.3.56
numpy==1.21.2
pandas==1.3.3
pytesseract==0.3.8
pdf2image==1.16.0
Pillow==8.3.2
python-socketio==5.4.0
openpyxl==3.0.9
